Only fetch an online UUID in online mode
The previous code would get an online UUID even in offline mode what
breaks plugins if the player joins.

Example:
You want to store data for player "Test"  who never joined. An online UUID is created and you save it using that UUID.

The player Test joins with an offline UUID but that will not match the online UUID of the saved data.